nutanix/python/requirements.txt
ntnx-networking-py-client
ntnx-prism-py-client
ntnx-vmm-py-client
python-dotenv